feat #157 데모 앱 추가
