+++ Projects/Features/MainFeature/Demo/Sources/MockLiveStreamViewControllerFactory.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+import LiveStreamFeatureInterface
+
+public struct MockLiveStreamViewControllerFractoryImpl: LiveStreamViewControllerFactory {
+    public init() { }
+    
+    public func make() -> UIViewController {
+        let viewModel = MockLiveStreamViewModel()
+        return MockLiveStreamViewController(viewModel: viewModel)
+    }
+}
