+++ Projects/Features/MainFeature/Sources/BroadcastCollectionViewController.swift
@@ -116,8 +116,7 @@ public class BroadcastCollectionViewController: BaseViewController<BroadcastColl
 // MARK: - CollectionView Delegate
 extension BroadcastCollectionViewController: UICollectionViewDelegate {
     public func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let viewModel = SampleLiveStreamViewModel()
-        let viewController = SampleLiveStreamViewController(viewModel: viewModel)
+        guard let viewController = factory?.make() else { return }
         viewController.modalPresentationStyle = .overCurrentContext
         viewController.transitioningDelegate = transitioning
         present(viewController, animated: true, completion: nil)
